Extract Hero component and organize code
