+++ docs/components/Homepage.module.css
@@ -1,186 +1,3 @@
-.hero {
-  position: relative;
-}
-
-.heroBg {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  background-image: linear-gradient(
-    180deg,
-    rgba(229, 231, 235, 0),
-    rgba(229, 231, 235, 0.7)
-  );
-
-  :global(.dark) & {
-    background-image: linear-gradient(
-      180deg,
-      rgba(38, 38, 38, 0.1),
-      rgba(38, 38, 38, 0.7)
-    );
-  }
-}
-
-.heroTiles {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  z-index: 1;
-  background-image: url('./heroTile.svg');
-  background-position: calc(50% - 35.5px) bottom;
-  background-repeat: repeat;
-  background-size: 142px 71px;
-
-  :global(.dark) & {
-    background-image: url('./heroTileDark.svg');
-  }
-}
-
-.heroContent {
-  margin: 0 auto;
-  position: relative;
-  z-index: 2;
-  padding-bottom: 142px;
-  padding-left: max(env(safe-area-inset-left), 1.5rem);
-  padding-right: max(env(safe-area-inset-right), 1.5rem);
-  max-width: 90rem;
-  text-align: center;
-}
-
-.headline {
-  margin-top: 4rem;
-  display: inline-flex;
-  background-image: linear-gradient(146deg, #000, #5e6265);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  background-clip: text;
-  font-size: 3.125rem;
-  font-size: min(4rem, max(8vw, 2.8rem));
-  font-weight: 700;
-  font-feature-settings: initial;
-  line-height: 1.05;
-  letter-spacing: -0.12rem;
-
-  :global(.dark) & {
-    background-image: linear-gradient(146deg, #fff, #81878b);
-  }
-}
-
-.subtitle {
-  margin-top: 1.25em;
-  color: hsl(204 20% 40%);
-  font-size: 1.3rem;
-  font-size: min(1.3rem, max(3.5vw, 1.1rem));
-  font-feature-settings: initial;
-  font-weight: 400;
-  line-height: 1.6;
-
-  :global(.dark) & {
-    color: hsl(204, 12%, 50%);
-  }
-}
-
-.actions {
-  margin-top: 1.6em;
-  margin-bottom: 1.4em;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  font-size: 1.3rem;
-  font-size: min(1.3rem, max(3.5vw, 1.1rem));
-  font-weight: 500;
-}
-
-a.cta {
-  margin-right: 0.75rem;
-  padding: 0.75rem 1.5rem;
-  border-radius: 6px;
-  box-shadow: 0 1px 2px #00295738;
-  background: linear-gradient(to bottom, #238aff, #0077ff);
-  color: white;
-  text-shadow: 0 1px 1px #00387838;
-  text-decoration: none;
-  white-space: nowrap;
-  user-select: none;
-  transition: all 0.2s ease;
-
-  span {
-    display: inline-block;
-    transition: all 0.2s ease;
-  }
-
-  &:hover {
-    box-shadow: 0 5px 30px -10px #0078ffab;
-    filter: brightness(1.05);
-
-    span {
-      transform: translateX(3px);
-    }
-  }
-
-  &:active {
-    box-shadow: 0 1px 3px #00295738;
-    filter: brightness(0.95);
-
-    span {
-      transform: translateX(5px);
-    }
-  }
-
-  &:focus-visible {
-    outline: 2px solid
-      hsl(var(--nextra-primary-hue) var(--nextra-primary-saturation) 77%);
-    outline-offset: 2px;
-
-    span {
-      transform: translateX(3px);
-    }
-  }
-}
-
-a.secondaryAction {
-  padding: 0.75rem 1.5rem;
-  border-radius: 6px;
-  box-shadow: hsl(204 10% 75%) 0 0 0 1px inset;
-  background: rgba(255, 255, 255, 0.7);
-  color: hsl(204 20% 30%);
-  text-decoration: none;
-  white-space: nowrap;
-  user-select: none;
-  transition: all 0.2s ease;
-
-  &:hover {
-    box-shadow: 0 5px 22px -10px hsl(204 45% 20%);
-    box-shadow: hsl(204 20% 50%) 0 0 0 1px inset;
-  }
-
-  &:active {
-    box-shadow: hsl(204 10% 75%) 0 0 0 1px inset;
-  }
-
-  &:focus-visible {
-    outline: 2px solid
-      hsl(var(--nextra-primary-hue) var(--nextra-primary-saturation) 77%);
-    outline-offset: 2px;
-  }
-
-  :global(.dark) & {
-    background: rgba(17, 17, 17, 0.7);
-    color: hsl(204 10% 70%);
-    box-shadow: hsl(204 10% 25%) 0 0 0 1px inset;
-    text-shadow: 0 1px 1px #000;
-
-    &:hover {
-      background: rgba(17, 17, 17, 1);
-      box-shadow: hsl(204 10% 50%) 0 0 0 1px inset;
-    }
-
-    &:active {
-      box-shadow: hsl(204 10% 25%) 0 0 0 1px inset;
-    }
-  }
-}
-
 .content {
   border-bottom: 1px solid rgb(229, 231, 235);
   background-color: #f3f4f6;
@@ -191,7 +8,7 @@ a.secondaryAction {
   }
 }
 
-a.demoLink {
+a.demoPreview {
   margin: 0 auto;
   margin-bottom: -143px;
   display: block;
@@ -199,60 +16,12 @@ a.demoLink {
   z-index: 2;
   top: -143px;
   max-width: 1208px;
-}
 
-img.demo {
-  display: block;
-  width: 100%;
-  box-sizing: border-box;
-  max-width: 1208px;
-  border: 1px solid rgba(55, 55, 55, 1);
-}
-
-.links {
-  margin-bottom: 3em;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-a.link {
-  margin-right: 0.25em;
-  padding: 0 0.5em;
-  color: hsl(204 20% 50%);
-  font-size: 1.2rem;
-  font-size: min(1.2rem, max(3.5vw, 1rem));
-  font-weight: 400;
-  line-height: 1.5em;
-  white-space: nowrap;
-
-  strong {
-    color: hsl(204 20% 30%);
-    font-weight: 400;
-    transition: all 0.2s ease;
-  }
-
-  &:last-child {
-    margin-right: 0;
-  }
-
-  &:hover {
-    strong {
-      color: hsl(204 20% 20%);
-    }
-  }
-
-  :global(.dark) & {
-    color: hsl(204, 12%, 40%);
-
-    strong {
-      color: hsl(204, 12%, 60%);
-    }
-
-    &:hover {
-      strong {
-        color: hsl(204, 12%, 70%);
-      }
-    }
+  img {
+    display: block;
+    width: 100%;
+    box-sizing: border-box;
+    max-width: 1208px;
+    border: 1px solid rgba(55, 55, 55, 1);
   }
 }
