+++ docs/images/logo/cosmonaut-black.png
