+++ docs/images/logo/cosmonaut-white.png
