+++ docs/next.config.js
@@ -11,13 +11,16 @@ export default withNextra({
   output: 'export',
   images: { unoptimized: true },
   webpack(config) {
-    const allowedSvgRegex = /components\/icons\/.+\.svg$/;
+    const allowedSvgRegex = /\/components\/svg\/.+\.svg$/;
 
     const fileLoaderRule = config.module.rules.find(rule => {
       return rule.test instanceof RegExp && rule.test.test('.svg');
     });
     fileLoaderRule.exclude = allowedSvgRegex;
 
+    // These rule allows importing SVGs as React components,
+    // but breaks the ability to use SVGs as background images in CSS,
+    // which Next.js handles by default.
     config.module.rules.push({
       test: allowedSvgRegex,
       use: ['@svgr/webpack'],
