+++ docs/pages/docs/index.mdx
@@ -1,5 +1,5 @@
 import { Cards, Card } from 'nextra/components';
-import StarIcon from '../../components/icons/stars.svg';
+import StarIcon from '../../components/svg/icons/stars.svg';
 
 # Introduction
